Safari-611.1.10/LayoutTests/inspector/unit-tests/inspector-test-dispatch-event-to-frontend-expected.txt


== Running test suite: InspectorTest.TestPage.dispatchEventToFrontend
-- Running test case: InspectorTest.TestPage.dispatchEventToFrontend
PASS: Received Event
PASS: Event name should be the event dispatched.
PASS: Event data should be included.
